unified dcrypting score button
